paging for fetching products

+++ backend/src/main/java/com/pcc/pc_configurator/repositories/CaseRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.relational.core.mapping.Table;
 import org.springframework.data.repository.CrudRepository;
 
-public interface CaseRepository extends JpaRepository<Case_,Integer> {
+public interface CaseRepository extends JpaRepository<Case_,Long> {
 }
